+++ src/main/java/com/springboot/publicplace/service/TeamJoinAdminService.java
@@ -1,0 +1,18 @@
+package com.springboot.publicplace.service;
+
+import com.springboot.publicplace.dto.ResultDto;
+import com.springboot.publicplace.dto.response.TeamJoinDetailResponseDto;
+import com.springboot.publicplace.dto.response.TeamJoinResponseDto;
+
+import javax.servlet.http.HttpServletRequest;
+import java.util.List;
+
+public interface TeamJoinAdminService {
+    List<TeamJoinResponseDto> getJoinRequestList(Long teamId, HttpServletRequest servletRequest);
+
+    TeamJoinDetailResponseDto getJoinRequestDetail(Long requestId, HttpServletRequest servletRequest);
+
+    ResultDto approveJoinRequest(Long requestId, HttpServletRequest servletRequest);
+
+    ResultDto rejectJoinRequest(Long requestId, HttpServletRequest servletRequest);
+}
