+++ src/main/java/com/springboot/publicplace/service/TeamJoinService.java
@@ -8,17 +8,8 @@ import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 public interface TeamJoinService {
-
     ResultDto joinTeam(Long teamId, HttpServletRequest servletRequest, String joinReason);
 
     ResultDto leaveTeam(Long teamId, HttpServletRequest servletRequest);
-
-    List<TeamJoinResponseDto> getJoinRequestList(Long teamId, HttpServletRequest servletRequest);
-
-    TeamJoinDetailResponseDto getJoinRequestDetail(Long requestId, HttpServletRequest servletRequest);
-
-    ResultDto approveJoinRequest(Long requestId, HttpServletRequest servletRequest);
-
-    ResultDto rejectJoinRequest(Long requestId, HttpServletRequest servletRequest);
 
 }
